Containers/uplink/requirements.txt
aiohttp == 3.7.4.post0
google-cloud-pubsub
asyncio
requests